feat: 提取输入工具栏为单独组件,添加事件总线
