+++ src/views/Home/Chat/components/ChatBox/MsgInput/styles.scss
@@ -51,7 +51,7 @@
   box-shadow: 0 0 8px 0 rgba(0, 0, 0, 10%);
   transform: translateY(-100%);
 
-  .person-warpper {
+  .person-wrapper {
     max-height: 250px;
     padding-right: 2px;
     overflow: auto;
